tweak max hunger and thirst, food on map and food used in reproduction reformat decisional code and reorder critter priorities remove gender and sexual reproduction reduce enviro size to 8 with dubious results fix bugs with size change change worldgen to simplex noise with circular filter, still wip edit saltwater to be drinkable for now change geneid system in favor of genelibrary, which actually works rebalance biomes for new worldgen fix enviro assignment problems add actions system to monitor critter behaviour

+++ src/cells/Puddle.java
@@ -18,7 +18,7 @@ public class Puddle extends LowWater {
     @Override
     public void tick() {
         Random r = enviro.getR ();
-        if ((r.nextInt () * enviro.getTemperature ()) % 1000 == 0 || CellList.valueOf (previous.getType ()).isPermeable () && r.nextInt () % 4 == 0) {
+        if ((r.nextInt () * (int) enviro.getTemperature ()) % 1000 == 0 || CellList.valueOf (previous.getType ()).isPermeable () && r.nextInt () % 4 == 0) {
             enviro.replaceWith (this, previous);
         }
     }
